+++ src/states/weapon_states.cpp
@@ -1,12 +1,12 @@
 #include "weapon_states.h"
-#include "../weapon_state_machine.h"
 #include "../weapon_manager.h"
+
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 ///////////////////////////////// Weapon Idle State ///////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 WeaponIdleState::WeaponIdleState(const WeaponStateData& weaponStateData)
-  : State(static_cast<int8_t>(WeaponStates::IDLE)), m_WeaponManager(weaponStateData.weaponManager),
+  : State(static_cast<int>(WeaponStates::IDLE)), m_WeaponManager(weaponStateData.weaponManager),
     m_WeaponStateMachine(weaponStateData.weaponStateMachine),
     m_WeaponStateContext(m_WeaponManager->get_weapon_state_ctx())
 {
@@ -17,12 +17,12 @@ void WeaponIdleState::_handle_input(const Ref<InputEvent>& event)
   if(Input::get_singleton()->is_action_just_pressed("shoot_weapon"))
   {
     m_WeaponStateContext.IsKeyPressed = true;
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::SHOOT));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::SHOOT));
   }
 
   if(Input::get_singleton()->is_action_just_pressed("reload_weapon"))
   {
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::RELOAD));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::RELOAD));
   }
   
 }
@@ -48,7 +48,7 @@ void WeaponIdleState::_exit()
 ///////////////////////////////// Weapon Equip State ///////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 WeaponEquipState::WeaponEquipState(const WeaponStateData& weaponStateData)
-  : State(static_cast<int8_t>(WeaponStates::EQUIP)), m_WeaponManager(weaponStateData.weaponManager),
+  : State(static_cast<int>(WeaponStates::EQUIP)), m_WeaponManager(weaponStateData.weaponManager),
     m_WeaponStateMachine(weaponStateData.weaponStateMachine)
 {
 }
@@ -57,7 +57,7 @@ void WeaponEquipState::_handle_input(const Ref<InputEvent>& event)
 {
   if(Input::get_singleton()->is_action_just_pressed("shoot_weapon"))
   {
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::SHOOT));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::SHOOT));
   }
 }
 
@@ -74,7 +74,7 @@ void WeaponEquipState::_enter()
 
 void WeaponEquipState::_update(double delta)
 {
-  m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::IDLE));
+  m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::IDLE));
 }
 
 
@@ -86,7 +86,7 @@ void WeaponEquipState::_exit()
 ///////////////////////////////// Weapon Shoot State ///////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 WeaponShootState::WeaponShootState(const WeaponStateData& weaponStateData)
-  : State(static_cast<int8_t>(WeaponStates::SHOOT)), m_WeaponManager(weaponStateData.weaponManager),
+  : State(static_cast<int>(WeaponStates::SHOOT)), m_WeaponManager(weaponStateData.weaponManager),
     m_WeaponStateMachine(weaponStateData.weaponStateMachine),
     m_WeaponStateContext(m_WeaponManager->get_weapon_state_ctx())
 {
@@ -96,7 +96,7 @@ void WeaponShootState::_handle_input(const Ref<InputEvent>& event)
 {
   if(Input::get_singleton()->is_action_just_pressed("reload_weapon"))
   {
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::RELOAD));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::RELOAD));
   }
 
 }
@@ -110,7 +110,7 @@ void WeaponShootState::_update(double delta)
 {
   if(Input::get_singleton()->is_action_just_pressed("shoot_weapon") && m_WeaponManager->get_current_weapon_ammo() == 0 && m_CurrentWeapon->get_auto_reload())
   {
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::RELOAD));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::RELOAD));
   }
 
   m_WeaponManager->_shoot_weapon(delta);
@@ -118,7 +118,7 @@ void WeaponShootState::_update(double delta)
   if(m_WeaponStateContext.ShootTimeBeforeIdle <= 0.0f &&
      m_WeaponStateContext.IsKeyPressed == false && m_WeaponStateContext.IsKeyHeld == false)
   {
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::IDLE));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::IDLE));
   }
 }
 
@@ -131,7 +131,7 @@ void WeaponShootState::_exit()
 ///////////////////////////////// Weapon Reload State ///////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 WeaponReloadState::WeaponReloadState(const WeaponStateData& weaponStateData)
-  : State(static_cast<int8_t>(WeaponStates::RELOAD)), m_WeaponManager(weaponStateData.weaponManager),
+  : State(static_cast<int>(WeaponStates::RELOAD)), m_WeaponManager(weaponStateData.weaponManager),
     m_WeaponStateMachine(weaponStateData.weaponStateMachine),
     m_WeaponStateContext(m_WeaponManager->get_weapon_state_ctx())
 {
@@ -156,14 +156,14 @@ void WeaponReloadState::_update(double delta)
 {
   if(Input::get_singleton()->is_action_just_pressed("shoot_weapon"))
   {
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::SHOOT));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::SHOOT));
   }
   
   m_WeaponManager->_reload_weapon();
 
   if(m_WeaponStateContext.IsReloading == false)
   {
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::IDLE));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::IDLE));
   }
 }
 
@@ -176,7 +176,7 @@ void WeaponReloadState::_exit()
 ///////////////////////////////// Weapon Unequip State ///////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 WeaponUnequipState::WeaponUnequipState(const WeaponStateData& weaponStateData)
-  : State(static_cast<int8_t>(WeaponStates::UNEQUIP)), m_WeaponStateMachine(weaponStateData.weaponStateMachine),
+  : State(static_cast<int>(WeaponStates::UNEQUIP)), m_WeaponStateMachine(weaponStateData.weaponStateMachine),
     m_WeaponManager(weaponStateData.weaponManager), m_WeaponStateContext(m_WeaponManager->get_weapon_state_ctx())
 {
 }
@@ -202,7 +202,7 @@ void WeaponUnequipState::_update(double delta)
 
   if(m_WeaponStateContext.IsUnequipped)
   {
-    m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::IDLE));
+    m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::IDLE));
   }
 }
 
@@ -215,7 +215,7 @@ void WeaponUnequipState::_exit()
 ///////////////////////////////// Weapon Switch State ///////////////////////////////////////
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 WeaponSwitchState::WeaponSwitchState(const WeaponStateData& weaponStateData)
-  : State(static_cast<int8_t>(WeaponStates::WEAPON_SWITCH)), m_WeaponManager(weaponStateData.weaponManager),
+  : State(static_cast<int>(WeaponStates::WEAPON_SWITCH)), m_WeaponManager(weaponStateData.weaponManager),
     m_WeaponStateMachine(weaponStateData.weaponStateMachine)
 {
 }
@@ -238,7 +238,7 @@ void WeaponSwitchState::_enter()
 
 void WeaponSwitchState::_update(double delta)
 {
-  m_WeaponStateMachine->_change_state(static_cast<int8_t>(WeaponStates::EQUIP));
+  m_WeaponStateMachine->_change_state(static_cast<int>(WeaponStates::EQUIP));
 }
 
 void WeaponSwitchState::_exit()
